@@ -6,15 +6,13 @@ import colorsys
 import contextlib
 import re
 import sys
+from collections.abc import Iterable, Sequence
 from typing import (
     TYPE_CHECKING,
     Any,
     Callable,
-    Iterable,
-    List,
     Literal,
     NamedTuple,
-    Sequence,
     SupportsFloat,
     Union,
     cast,
@@ -53,7 +51,7 @@ ColorLike: TypeAlias = Union[
     RGBTuple,  # 3-tuple of all ints or all floats
     RGBATuple,  # 4-tuple of all floats, or 3 ints and 1 float
     np.ndarray,  # 3- or 4-element rgb(a) vector
-    List[Union[float, int]],  # 3- or 4-element rgb(a) vector
+    list[Union[float, int]],  # 3- or 4-element rgb(a) vector
     "Color",  # another color object
 ]
 """Data types that can be cast to a [cmap.Color][] instance.
